requirements.txt
Flask==1.1.2
pandas==1.1.5
streamlit==0.78.0
spacy==3.0.3
Wikipedia_API==0.5.4
altair==4.1.0
PyJWT==2.0.1
gunicorn==20.0.4
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0.tar.gz#egg=en_core_web_sm
